New version of boOX.

+++ boOX-y/src/glucose/Alloc.h
@@ -62,11 +62,11 @@ class RegionAllocator
     void     free      (int size)    { wasted_ += size; }
 
     // Deref, Load Effective Address (LEA), Inverse of LEA (AEL):
-    T&       operator[](Ref r)       { assert(r >= 0 && r < sz); return memory[r]; }
-    const T& operator[](Ref r) const { assert(r >= 0 && r < sz); return memory[r]; }
+    T&       operator[](Ref r)       { assert((int)r >= 0 && r < sz); return memory[r]; }
+    const T& operator[](Ref r) const { assert((int)r >= 0 && r < sz); return memory[r]; }
 
-    T*       lea       (Ref r)       { assert(r >= 0 && r < sz); return &memory[r]; }
-    const T* lea       (Ref r) const { assert(r >= 0 && r < sz); return &memory[r]; }
+    T*       lea       (Ref r)       { assert((int)r >= 0 && r < sz); return &memory[r]; }
+    const T* lea       (Ref r) const { assert((int)r >= 0 && r < sz); return &memory[r]; }
     Ref      ael       (const T* t)  { assert((void*)t >= (void*)&memory[0] && (void*)t < (void*)&memory[sz-1]);
         return  (Ref)(t - &memory[0]); }
